`2.0.0` release. (#82)
* Adjustments

* Oops

* Add support for audio flush.

This fixes some video stopping issues.

* if `lime_openal` is not defined, skip audio playback.

* Remove `LibVLC_Audio_Output_T` from da types.

* Breaking Change: Removed `getCalculatedVolume`.

* Remove `autoPause` and `autoVolumeHandle` and remake `autoResizeMode`.

The main reason behind the removal of these varibales is that they arent really a necesity to be used anymore, for example, the autoPause behaviour in the videos has been fixed for a while and the `autoVolumeHandle` has been dead for a long time.

* Fix compile

* When you touch the video, it can now pause.

* Fix null safety

* Fix sample compiling when `HXVLC_ENABLE_STATS` is defined.

* Implemented `volumeAdjust` and fixed volume getting/setting issues.

* Add `lime_openal` within the volume setters/getters.

* Leave `audioSetVolume` function empty.

* Multiple channel support audio.

* Oops

* Don`t use `tryAcquire`.

* Don't include the `lua` folder by default.

* Update haxelib an license

* Fix for `autoPause`.

* Fix some stuff with multiple channel support.

* Fixed some small bugs

* Fix compile issues when `FLX_SOUND_SYSTEM ` is not defined.

* Unify android and ios jobs.

* Simplified desktop workflow os

* Oops

* Set the volume to `125` as default.

According to https://github.com/videolan/vlc/blob/9ab3f7da8b434fa5e348f5e1d679df66eef36511/modules/gui/qt/qt.cpp#L430, this is default volume in the vlc application, so we are gonna use that.

* Fixed some issues

* Add ios building scripts

* Add the ability to use custom libvlc static libs on ios.

* Some fixes

* Update Android and iOS libs.

* Oops

* This should be better.

* Fix

* .........

* Update Android and iOS libs.

* Create README.md